feat: migrate from deprecated Memory module to Wiki knowledge system
- Remove Memory module documentation and data structures (MemoryMeta, MemoryFile, MemoryQuery, MemoryHit) from docs, replaced with Wiki structures
- Delete obsolete Embedder trait implementation for TextEmbedder in storage
- Clean up Pipeline prepare_pipeline by removing three-phase memory loading step
- Update AgentSession hook registration to use AgentConfig-based approach
- Simplify event_store bindings and page_store SQL queries, remove unnecessary type casts
- Apply minor formatting improvements across wiki tools, lifecycle, and migrations modules

+++ gasket/storage/src/wiki/page_store.rs
@@ -80,7 +80,7 @@ impl WikiPageStore {
         .bind(page.frequency.to_string())
         .bind(page.access_count as i64)
         .bind(page.last_accessed.as_deref())
-        .bind(page.file_mtime as i64)
+        .bind(page.file_mtime)
         .execute(&self.pool)
         .await?;
         Ok(())
@@ -160,13 +160,11 @@ impl WikiPageStore {
 
     /// Update only the frequency of a page (used by decay).
     pub async fn update_frequency(&self, path: &str, frequency: Frequency) -> Result<bool> {
-        let result = sqlx::query(
-            "UPDATE wiki_pages SET frequency = ? WHERE path = ?"
-        )
-        .bind(frequency.to_string())
-        .bind(path)
-        .execute(&self.pool)
-        .await?;
+        let result = sqlx::query("UPDATE wiki_pages SET frequency = ? WHERE path = ?")
+            .bind(frequency.to_string())
+            .bind(path)
+            .execute(&self.pool)
+            .await?;
         Ok(result.rows_affected() > 0)
     }
 }
